azure-vnet-api/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic
azure-identity
azure-mgmt-network
sqlite-utils
